pyproject.toml
[project]
name = "ptk-repl-cli"
version = "0.1.0"
description = "A modern REPL CLI with state sharing based on prompt-toolkit + Pydantic"
authors = [
    {name = "Your Name", email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "colorama>=0.4.6",
    "paramiko>=3.4.0",
    "prompt-toolkit>=3.0.52",
    "pydantic>=2.10.0",
    "pyfiglet>=1.0.4",
    "pyyaml>=6.0.3",
    "questionary>=2.0.0",
    "types-colorama>=0.4.15.20250801",
    "types-paramiko>=4.0.0.20250822",
]

[project.scripts]
ptk_repl = "ptk_repl.cli:main"

[project.optional-dependencies]
dev = [
    "ruff>=0.9.0",
    "mypy>=1.14.0",
    "types-pyyaml>=6.0.12.20250915",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ptk_repl"]

[dependency-groups]
dev = [
    "ruff>=0.9.0",
    "mypy>=1.14.0",
    "types-pyyaml>=6.0.12.20250915",
    "pre-commit>=3.0.0",
]

[tool.uv]
default-groups = ["dev"]

# Pre-commit scripts (run with: uv run pre-commit)
[tool.scripts]
pre-commit-check = "bash -c 'echo \"🔍 Running pre-commit checks...\" && uv run ruff check src/ && uv run mypy src/ && echo \"✅ All checks passed!\"'"
format = "uv run ruff format src/"
check = "uv run ruff check src/"
typecheck = "uv run mypy src/"
lint = "uv run ruff check src/ && uv run mypy src/"

[tool.ruff]
target-version = "py312"
line-length = 100
indent-width = 4

[tool.ruff.lint]
# 启用推荐的规则
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
